[DOM-335] Refactor validation
